refactor(admin): extract WaitlistSection (UI only)
Move waitlist tab content to dedicated component. Remove unused
icon imports from App.jsx (ChevronLeft, ChevronRight, Trash2 now
imported directly in WaitlistSection and StatusSection).
