+++ src/admin/App.jsx
@@ -23,10 +23,6 @@ const getDeviceId = () => {
 
 // Import extracted components
 import {
-  // Icons (used in waitlist tab)
-  Trash2,
-  ChevronLeft,
-  ChevronRight,
   // UI Components
   HoverCard,
   QuickActionsMenu,
@@ -63,6 +59,7 @@ import { AnalyticsSection } from './tabs/AnalyticsSection';
 import { BlockingSection } from './tabs/BlockingSection';
 import { TabNavigation } from './tabs/TabNavigation';
 import { StatusSection } from './tabs/StatusSection';
+import { WaitlistSection } from './tabs/WaitlistSection';
 
 // Feature flag: use real AI assistant instead of mock
 const USE_REAL_AI = true;
@@ -895,58 +892,11 @@ const AdminPanelV2 = ({ onExit }) => {
             />
           )}
           {activeTab === 'waitlist' && (
-            <div className="bg-white rounded-lg shadow-sm p-6">
-              <h3 className="text-lg font-semibold text-gray-800 mb-4">
-                Waiting Groups ({waitingGroups.length})
-              </h3>
-
-              {waitingGroups.length === 0 ? (
-                <p className="text-gray-500 text-center py-8">No groups waiting</p>
-              ) : (
-                <div className="space-y-3">
-                  {waitingGroups.map((group, index) => (
-                    <div
-                      key={index}
-                      className="flex items-center justify-between p-4 bg-gray-50 rounded-lg"
-                    >
-                      <div>
-                        <p className="font-medium">
-                          Position {index + 1}: {(group.names || []).join(', ') || 'Unknown'}
-                        </p>
-                        <p className="text-sm text-gray-600">
-                          {(group.names || []).length} player
-                          {(group.names || []).length !== 1 ? 's' : ''}
-                        </p>
-                      </div>
-                      <div className="flex gap-2">
-                        {index > 0 && (
-                          <button
-                            onClick={() => moveInWaitlist(index, index - 1)}
-                            className="p-1 text-blue-600 hover:bg-blue-50 rounded"
-                          >
-                            <ChevronLeft size={20} />
-                          </button>
-                        )}
-                        {index < waitingGroups.length - 1 && (
-                          <button
-                            onClick={() => moveInWaitlist(index, index + 1)}
-                            className="p-1 text-blue-600 hover:bg-blue-50 rounded"
-                          >
-                            <ChevronRight size={20} />
-                          </button>
-                        )}
-                        <button
-                          onClick={() => removeFromWaitlist(index)}
-                          className="p-1 text-red-600 hover:bg-red-50 rounded"
-                        >
-                          <Trash2 size={20} />
-                        </button>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div>
+            <WaitlistSection
+              waitingGroups={waitingGroups}
+              moveInWaitlist={moveInWaitlist}
+              removeFromWaitlist={removeFromWaitlist}
+            />
           )}
           {activeTab === 'analytics' && (
             <AnalyticsSection backend={backend} AnalyticsDashboard={AnalyticsDashboard} />
